requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic
python-multipart
python-dotenv
requests
PyMuPDF
google-generativeai
numpy
python-pptx
Pillow
openpyxl
python-docx
pytesseract
httpx
faiss-cpu
scikit-learn
nltk
spacy
psutil
langdetect
bs4
html2text
